requirements.txt
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Django==5.0.6
django-appconf==1.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-yasg==1.21.7

mistune==3.0.2
mysql-client==0.0.1
mysqlclient==2.2.4


